- basic data structures and math is moved to "sources/base" - headers are restructured (used "forward declarations" wherever possible) - a couple of files is renamed to be in common style with others.
