+++ sources/base/Numbered.cpp
