+++ sources/base/Particle.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <cmath>
 #include <string>
 #include "Numbered.h"
