--- sources/utils.h
@@ -1,5 +1,0 @@
-#pragma once
-
-int max(int a, int b);
-
-int min(int a, int b);
